+++ packages/components/src/renderer/carousel.tsx
@@ -4,8 +4,8 @@ import Slider, { Settings } from 'react-slick'
 // Added those manually to remove the font dependencies which keeps showing 404 not found
 import '../../../css/slick-theme.css'
 import '../../../css/slick.css'
-import { ActionButton, ActionType } from '../../content-typings'
-import { MessageTypeHandlerProps } from '../../typings'
+import { ActionButton, ActionType } from '../content-typings'
+import { MessageTypeHandlerProps } from '../typings'
 
 export class Carousel extends React.Component<MessageTypeHandlerProps<'carousel'>, ICarouselState> {
   private ref = createRef<HTMLDivElement>()
